Move bouncing balls example to separate project
